Reduced lines of code, Mentioned HTML deployment, Enhanced data_extraction.R efficiency, Enhanced makefile.

+++ src/data-preparation/data_extraction.R
@@ -13,22 +13,31 @@ download_and_process_data <- function(url, file_name, delim = "\t", data_dir = "
   # Define the file path
   file_path <- here(data_dir, paste0(file_name, ".tsv.gz"))
   
-  # Download the file
-  download.file(url, file_path)
-  print(paste(file_name, "file downloaded"))
-  
-  # Read the compressed file
-  data <- read_delim(gzfile(file_path), delim = delim)
-  print(paste(file_name, "data loaded into dataframe"))
-  
-  # Save the data as .csv
+  # Define the CSV path
   csv_path <- here(data_dir, paste0(file_name, "_data.csv"))
-  write_csv(data, csv_path)
-  print(paste(file_name, "data saved as CSV at", csv_path))
+  
+  # Check if the CSV file already exists
+  if (file.exists(csv_path)) {
+    print(paste(file_name, "data already exists as CSV. Skipping download and processing."))
+  } else {
+    # Download the file
+    download.file(url, file_path)
+    print(paste(file_name, "file downloaded"))
+  
+    # Read the compressed file
+    data <- read_delim(gzfile(file_path), delim = delim)
+    print(paste(file_name, "data loaded into dataframe"))
+  
+    # Save the data as .csv
+    write_csv(data, csv_path)
+    print(paste(file_name, "data saved as CSV at", csv_path))
+  }
 }
 
 ### Input / Transformation / Output ###
 # Call the function for ratings, genre, and episode data
 download_and_process_data("https://datasets.imdbws.com/title.ratings.tsv.gz", "rating")
+Sys.sleep(2)
 download_and_process_data("https://datasets.imdbws.com/title.basics.tsv.gz", "genre")
+Sys.sleep(2)
 download_and_process_data("https://datasets.imdbws.com/title.episode.tsv.gz", "episode")
